# api.yaml
swagger: "2.0.0"
info:
  title: Academy APP
  description: Api Documentation for Academia Cyber Flow
  version: 1.0.0
  contact:
    email: choudharysidharth082000@gmail.com
  servers:
    - url: "http://localhost:3000"
      description: "Local Server Running"
    - url: "https://testcyberflow.herokuapp.com"
      description: "Main Server"
schemes: [http, https]
paths:
...
  /v1/course/filterCourse:
    get:
      tags:
        - name: Course
      summary: Get All Courses By filter
      description: APi to get all the Subjects by filter
      parameters:
        - name: searchString
          in: query
          type: string
        - name: status
          in: query
          type: string
          enum: [active, inactive]
        - name: boardID
          in: query
          type: string
        - name: subBoardID
          in: query
          type: string
        - name: subjectID
          in: query
          type: string
        - name: classesID
          in: query
          type: string
        - name: startDate
          in: query
          type: string
        - name: endDate
          in: query
          type: string
        - name: page
          in: query
          type: string
        - name: limit
          in: query
          type: string
        - name: sortByDate
          in: query
          type: string
          enum: [asc, dsc]
      responses:
        200:
          description: OK
        400:
          description: Invalid Credentials
        404:
          description: Not Found
        401:
          description: Invalid Credentials
        500:
          description: Internal Server Error
  /v1/course/getCourseById/{courseID}:
    get:
      tags:
        - name: Course
      summary: Get All Courses By courseID
      description: APi to get all the Courses
      parameters:
        - name: courseID
          in: path
          type: string
          required: true
      responses:
        200:
          description: OK
        400:
          description: Invalid Credentials
        404:
          description: Not Found
        401:
          description: Invalid Credentials
        500:
          description: Internal Server Error
...
  /v1/classes/filterClasses:
    get:
      tags:
        - name: Classes
      summary: Get All Classes By filter
      description: APi to get all the Classes by filter
      parameters:
        - name: searchString
          in: query
          type: string
        - name: status
          in: query
          type: string
          enum: [active, inactive]
        - name: startDate
          in: query
          type: string
        - name: endDate
          in: query
          type: string
        - name: page
          in: query
          type: string
        - name: limit
          in: query
          type: string
        - name: sortByDate
          in: query
          type: string
          enum: [asc, dsc]
      responses:
        200:
          description: OK
        400:
          description: Invalid Credentials
        404:
          description: Not Found
        401:
          description: Invalid Credentials
        500:
          description: Internal Server Error
  /v1/classes/getClassesById/{classesID}:
    get:
      tags:
        - name: Classes
      summary: Get All classes By classesID
      description: APi to get all the Classes
      parameters:
        - name: classesID
          in: path
          type: string
          required: true
      responses:
        200:
          description: OK
        400:
          description: Invalid Credentials
        404:
          description: Not Found
        401:
          description: Invalid Credentials
        500:
          description: Internal Server Error
...
  /v1/resource/createResource:
    post:
      tags:
        - name: Resource
      summary: Create Resource
      description: Api to Create Resource
      consumes: multipart/form-data
      produces: application/json
      parameters:
        - name: name
          in: formData
          type: string
          required: true
        - name: description
          in: formData
          type: string
          required: true
        - name: linkType
          in: formData
          type: string
          enum: [upload, link]
          required: true
        - name: linkString
          in: formData
          type: string
        - name: resourceFile
          in: formData
          type: file
          required: false
        - name: status
          in: formData
          type: string
          enum: [active, inactive]

      responses:
        200:
          description: OK
        400:
          description: Invalid Credentials
        404:
          description: Not Found
        401:
          description: Invalid Credentials
        500:
          description: Internal Server Error
  /v1/resource/editResource/{resourceID}:
    patch:
      tags:
        - name: Resource
      summary: Edit Resource
      description: Api to Edit Resource
      consumes: multipart/form-data
      produces: application/json
      parameters:
        - name: resourceID
          in: path
          type: string
          required: true
        - name: name
          in: formData
          type: string
        - name: description
          in: formData
          type: string
        - name: linkType
          in: formData
          type: string
          enum: [upload, link]
        - name: linkString
          in: formData
          type: string
        - name: resourceFile
          in: formData
          type: file
        - name: status
          in: formData
          type: string
          enum: [active, inactive]
      responses:
        200:
          description: OK
        400:
          description: Invalid Credentials
        404:
          description: Not Found
        401:
          description: Invalid Credentials
        500:
          description: Internal Server Error
  /v1/resource/deleteResourceById/{resourceID}:
    delete:
      tags:
        - name: Resource
      summary: Delete Resource
      description: Delete Resource by its id
      parameters:
        - name: resourceID
          in: path
          type: string
          required: true
      responses:
        200:
          description: OK
        400:
          description: Invalid Credentials
        404:
          description: Not Found
        401:
          description: Invalid Credentials
        500:
          description: Internal Server Error

          ## Topic Apis

  ## CMS Apis
...